% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSquirrels.R
\name{fetch_squirrels}
\alias{fetch_squirrels}
\title{Get Squirrel Data!}
\usage{
fetch_squirrels()
}
\value{
a tibble of squirrel data
}
\description{
Provided neatly as a tibble and downloaded
via the \code{RSocrata} package
}
\examples{
fetch_squirrels()
}
